[BE] FIX: 밴 당한 댓글 삭제 불가 제약 조건 추가

+++ backend/src/main/java/org/ftclub/cabinet/presentation/service/PresentationCommentService.java
@@ -127,6 +127,9 @@ public class PresentationCommentService {
 		if (!comment.getUser().getId().equals(dto.getUserId())) {
 			throw ExceptionStatus.PRESENTATION_COMMENT_NOT_AUTHORIZED.asServiceException();
 		}
+		if (comment.isBanned()) {
+			throw ExceptionStatus.PRESENTATION_COMMENT_BANNED.asServiceException();
+		}
 
 		comment.delete();
 	}
